+++ settings.php
@@ -225,16 +225,24 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/backgroundimage';
         $title = get_string('backgroundimage', 'theme_boost', null, true);
         $description = get_string('backgroundimage_desc', 'theme_boost', null, true);
+        $description .= get_string('backgroundimage_desc_note', 'theme_boost_union', null, true);
         $setting = new admin_setting_configstoredfile($name, $title, $description, 'backgroundimage');
         $setting->set_updatedcallback('theme_reset_all_caches');
         $tab->add($setting);
 
         // Replicate the Login Background image setting from theme_boost.
         $name = 'theme_boost_union/loginbackgroundimage';
-        $title = get_string('loginbackgroundimage', 'theme_boost', null, true);
-        $description = get_string('loginbackgroundimage_desc', 'theme_boost', null, true);
-        $setting = new admin_setting_configstoredfile($name, $title, $description, 'loginbackgroundimage');
+        $title = get_string('loginbackgroundimage', 'theme_boost_union', null, true);
+        $description = get_string('loginbackgroundimage_desc', 'theme_boost_union', null, true);
+        $setting = new admin_setting_configstoredfile($name, $title, $description, 'loginbackgroundimage', 0,
+                array('maxfiles' => 25, 'accepted_types' => 'web_image'));
         $setting->set_updatedcallback('theme_reset_all_caches');
+        $tab->add($setting);
+
+        $name = 'theme_boost_union/loginbackgroundimagetext';
+        $title = get_string('loginbackgroundimagetextsetting', 'theme_boost_union', null, true);
+        $description = get_string('loginbackgroundimagetextsetting_desc', 'theme_boost_union', null, true);
+        $setting = new admin_setting_configtextarea($name, $title, $description, null, PARAM_TEXT);
         $tab->add($setting);
 
         // Create brand colors heading.